More tuning and tweaks
* Override and upgrade Flyway to 9.22.3
  * include flyway-mysql dependency in repositories
* Adjust heap size for Elasticsearch (it's greedy)
  * update DEV_NOTES and provision.sh script
* Revert to Mysql 8.1 (both driver and Docker image used)
* Adjust support child modules' dependencies
  * config server gets embedded application.yml
  * discovery service gets embedded application.yml
* Minor adjustments to sh scripts

+++ destroy.sh
@@ -11,4 +11,9 @@ name=$1
 
 
 # Destroy a Docker Machine instance
-docker-machine rm $name
+if ! command -v docker-machine &> /dev/null
+then
+  docker-machine rm $name
+else
+  echo "Cannot execute destruction request, docker-machine is not installed!"
+fi
